Set http prefix to url in sharing mail
If the url specified in the recipient's structure is not prefixed by
either "http" or "https" the button generated in the invitation email
cannot be correctly handled.

This PR fixes this problem by prefixing it by the appropriate value,
depending on `instance.Scheme()`. If the url is already prefixed then
nothing is changed.