Add sharer structure to sharing
We need to be able to access the sharer's information from the sharing
document: this is what this PR adds.

The `sharer` structure is filled when a sharing request is received by a
recipient. It is composed of the id of the sharer's OAuth Client (at the
recipient's) and of the sharer's URL of her Cozy.

Tests where updated accordingly.