A recipient can receive a sharing.
This PR adds a route `/sharings/doc/:doctype/:id` that will handle the
documents sent by the sharer. Depending on the doctype either the
request is transmitted to `/data` or our custom handler (for the
creation of a file/dir with an id) is called.

This route was added because we need to modify the parent of an incoming
file/dir and set it to "Shared With Me" as we don't want to propagate
the entire hierarchy with a document.

We created a custom handler that heavily resembles that of files because
we need to be able to predefine the id of a file/dir for synchronisation
purposes. To do so we had to add two new functions to the interface of
the indexer and update the vfs implementation.

The list of changes is:
* consts.go: add `SharedWithMe` constant.
* share_data.go: the functions `SendDoc` and `SendFile` now call the
  correct route on the recipient's Cozy.
* sharings_test.go: it is no longer possible to check on the recipient's
  test server that the sharing went well. The route called to share a
  document is in web/sharings and we can't add it to the test server
  handler as it would create a dependency cycle. This is a *temporary*
  patch as fixing everything is out of the scope of this PR.
* couchdb_indexer.go: add the functions `CreateNamedFileDoc` and
  `CreateNamedDirDoc` to be able to create a file/dir with an id.
* vfs.go: update the interface of the indexer to include the functions
  `CreateNamedFileDoc` and `CreateNamedDirDoc`.
* vfsafera/impl.go: When creating a file or a directory if an ID is
  already present in the document then the corresponding `Named*Doc`
  function is called.
* vfsswift/impl.go: (same as above).
* web/files/files.go: to ease the creation of a file/dir with a given id
  the function `FileDocFromReq` was turned public so that I can access it
  from the package Sharings.
* web/sharings/files.go: creation of a file or dir with an id, plus
  creation of the dir "Shared With Me" if it doesn't exist.
* web/sharings/sharings.go: add the route `/sharings/doc/:doctype/:id`
  and its handler function `receiveDocument`.
* web/sharings/sharings_test.go: add tests for the newly added route.