Share and update JSON documents
With a previous PR two regressions were introduced: it was no longer
possible to share a JSON, and its updates were not propagated. This PR
fixes those issues and enhance the propagation of updates as it is now
possible for a sharer to propagate the deletion of a shared document.