Share files updates
With this PR when a file is shared in Master-Slave it will be updated at
the recipients' accordingly to what the sharer does.

A problem exists with the `UpdatedAt` field which is not automatically
updated when a file is, so the test in the `vfs` package was replaced by
an error log.