Handle wrong tag detection
"git --tags" report a random tag (seems the first) if multiple tags are on the same commit.
This breaks the Debian package build (set a tag but obtain another version binary) in case of multiples rebuild without upstream changes.
This patch allows to force the version string and to not rely on tags.