feat: Get permissions for Cozy-to-Cozy sharings
Following /permissions/doctype/:doctype two routes were added to fetch
the list of permissions associated to Cozy-to-Cozy sharings.  The first
one fetches the list of permissions for sharings for which the user is a
recipient ("sharedWithMe").
The second one fetches the list of permissions for sharings for which
the user is the sharer ("sharedWithOthers").

For clarity the route /permissions/doctype/:doctype was renamed to
/permissions/doctype/:doctype/sharedByLink.