Fix shutdown of redis broker
The redis broker is waiting in the shutdown for all the workers to stop
polling redis for new jobs. Before this commit, in the list of workers,
there were also workers with concurrency at 0, and for those workers, we
didn't start the polling. So, on a server where some workers were
disabled, we were forced to wait until the context timeout on shutdown
even if there was no more activity.