Fix Swift encoding (#1645)
Disable Accept-Encoding for requests to Swift. For files with a gzipped
size inferior to the real size, a reverse-proxy between the stack and
Swift can gzip the response, and when it is the case, the stack may
return a 500 because it can no longer infer the file size from the
Content-Length, as http.ServeContent does.