Improve the MIME-Type handling (#1793)
The magic package was not maintained and some types like webp images
were missing. So, I've replaced it with an external library, filetype.

And, I think we were using a Content-Type with a charset at some places
where a type and subtype, but no charset, would be better. I've changed
that too.