Load the dynamic assets in background
This commit changes the loading of the assets and its behaviour when
failing to do so. Previously, the stack will block for about one minute
to retry several times to fetch the dynamic assets, and exit with an
error after that. It means that one missing dynamic assets was enough to
make the cozy-stack unavailable of the whole production environment.
Now, the cozy-stack serves normally the request (with a fallback for the
missing assets), and just logs the error for loading the assets.

Fix #1802