Extract the build variables from pkg/config
The pkg/config is quite big and is a dependency of the client package.
To limit what is imported by the client, this commit split pkg/config:

- the build variables remain in pkg/config (exported as build)
- other things go to pkg/config/config (exported as config).

Ideally, it would have been cleaner to move the build variables to their
own package (like pkg/config/build), but as these variables are injected
by some scripts (scripts/release.sh but also externals scripts), it
seems safer to do it in the opposite way.