Allow clients to override the created_at for files
When a client uploads a file or create a new directory, it can set the
updated_at via an HTTP header (Date), but it can also set the created_at
via the CreatedAt parameter in the query-string. Both the created_at and
updated_at dates are controlled by the client. The server keeps its date
in cozyMetadata.