Introduce a new Swift layout for the VFS
The new Swift layout V3 (but numbered 2 in code, as indexed from 0) has
been added to the VFS. It will help to manage file versioning. New Cozy
instances will use it by default from now. But for old instances, a
migration will be needed, but nothing has been scripted to do that yet.