Add KdfIterations field to the instance model
We want to use the same password from the user to log in the cozy and to
encrypt data on the clients. To avoid sending the encryption key to the
server, we will derive two keys from the master password: one for the
login, and one for the encryption.

With this commit, a first step is made in this direction, by allowing to
store the parameters for deriving the login key from the master password
in the instance model.