Fix the realtime event from CouchDB updates (bis)
When using couchdb.UpdateDoc, the old document was a copy of the new
document, but UnmarshalJson does not reset the old values. So, if a
field was added to the document, the realtime saw this field in the old
document.

With this fix, the old document is created as an empty object of the
same type of the new document, and, as such, won't have the fields
added to the new document.

It is particular helpful for triggers with a changed value, as the
trigger didn't push a job when the selected field was added before this
fix.