Use some predefined credentials in dev mode
The vault must have some credentials to work. They are configured via
the configuration file or the command line flag. But, for development,
using them is not straight forward, and with this commit, a set of
predefined credentials will be used in that case (and only if the user
has not set credentials via its config).