Fix behavior when an instance is blocked for TOS (#2111)
If an instance is blocked because the user hasn't signed the terms of
services, the requests to this instance are rejected. They were always
redirected to the manager, but now, only the requests for an HTML page
are redirected. The requests for a JSON API are rejected with 402
Payment required, and a JSON message.