Improve the Swift container deletion (#2143)
When deleting a Swift container, we have to first delete all the objects
inside it, and then delete the container its-self. Swift behavior on
consistency makes possible that the call to delete the container can
fail just after we have deleted all the objects. This commit adds a
retry loop to work around this limitation.