Update linked app on OAuth refresh (#2225)
When an OAuth application is linked to a web app, its permission are
taken from the web app manifest. If the web app is updated to add new
permissions, the OAuth application may expect that, after some time, it
can use the new permissions. For people using only the OAuth
application, but not the web app, the manifest can be never updated with
the new permissions. This commit tries to avoid this issue by updating
the linked web appl when the OAuth application refreshs its token. So,
if someone uses an OAuth application 7 days after the linked web app has
been updated on the registry, their Cozy instance will have the new
version of the application, and the OAuth app can use the new
permissions.