Compute the bucket size for exports (#2230)
When a user asks to export its data with the files, it can take a lot of
space. For practical reasons, we split the archive in several parts (the
buckets). The size of a bucket was limited to 100MB before this commit.
It means that for large instance, a user must have to download hundreds
or thousands of buckets to export the whole. It has been changed to a
dynamic rule. For 100MB or less, it is still a single bucket. For more,
the bucket size will increase in the same proportion as the number of
buckets. So, if an instance takes 4x more space than another, its export
will use 2x more buckets and the buckets will be 2x larger. For an
instance with 1TB, it means 100 buckets of 10GB.