Allow to use wildcard for the permissions
It is now possible to add the suffix .* to a doctype in a permission to
give access to this doctype and its sub-doctypes. For example, a
permission on io.cozy.bank.* will give access to io.cozy.bank,
io.cozy.bank.accounts, io.cozy.bank.settings, etc.

It can be useful to avoid asking the user for new permissions when an
application is adding new doctypes on a domain.