Rework opening shared notes
A note can be opened on the same Cozy instance as it has been created by
the owner of its instance (easy). Or, the note can be inside a directory
shared by link, and the visitor can also open the note with the same
sharecode.

And there are the notes inside a Cozy-to-Cozy sharing. The note should
be opened on the instance where it has been created. The sharecode must
gives read-only permission if the person that tries to open the note is
marked as read-only in the sharing, or if the sharing rules forbid
writing the note. Else, it must be read-write.

Note that we have to take into account the star model used for sharing.
Two recipients instances can't talk directly as they have no common
authentication proof, the requests must be proxified by the sharer
instance.