Make recreateParent a bit more robust
The recreateParent function of sharing tries to recreate a missing
directory. As it makes some network calls, it can be slow and the
missing directory can be created in the background by another request to
the stack. So, when it fails with a os.ErrExist error, the stack can try
to fetch the document from the database.