Improve notes handling in shared directory
When a note is in a shared directory, the stack should avoid to modify
the content and rename the note in a single CouchDB revision. It doesn't
play nicely with the sharing, as the stack has to create a new revision
on the recipients' instances. And if there are several instances, the
new revisions can be different, leading to conflicts.

To fix that, the stack now makes a first CouchDB revision with the
renaming, and only after that, writes the new content to the VFS.