Change the rule for conflict on same path
Previously, the rule was based on a comparison of the identifier. Now,
the rule is that the owner instance wins. It helps to minimize the
number of exchanges needed by the sharing to converge. As such, it can
avoid some unecessary additional conflicts when the sharing has a lot of
members.

The rule for naming the file in conflict is also changed to add a (2)
suffix (instead of - conflict - number), as it is more understable for
the users.