Add an option to skip OAuth state on token (#2496)
When a konnector uses an OAuth provider, the stack is responsible to
manage the OAuth dance. Some providers require the state on the access
token endpoint, and some others throw en error if the state is given.
This commit adds an option to the account_types, named
skip_state_on_token to not include the state on this particular request.