Improve permissions checking for notes and shortcuts (#2479)
When a note or a shortcut is created via their special API, the stack
checks the permission to write on the VFS. Before this commit, only a
permission on the whole io.cozy.files doctype was authorized. Now, we
also accept a permission on the directory where the note or shortcut
will be created. It is useful when a directory is shared by link for
example.