Add lock on sharing reception
We used to have a lock when sending shared documents but not on their
reception. This could cause CouchDB conflicts when an instance receives
at the same time changes on the same document from different instances.

To illustrate a possible scenario of such conflict, let's assume a shared directory with the revisions list `{1-aa,
2-aa}` between the instances A, B and C.

At the same time, instance A receives changes from B and C, respectively with the revisions lists `{1-aa, 2-bb}` and
`{1-aa, 2-cc, 3-cc}`.

Changes from B are applied: the revision `2-bb` conflicts with the local `2-aa`, and
is solved by generating a new local update, producing the revisions list
`{1-aa, 2-aa, 3-aa}`.

Simultaneously, the changes from C are applied: the last local revision
is still seen as `2-aa` and the following revision list is forced: `{1-aa,
2-aa, 3-cc}`. Then, a CouchDB conflict is produced between `3-aa` and `3-cc`.