Force trashing file (#2693)
When a file is requested to be put in the trash, the stack looks if it
is not already in the trash. It was done via checking if its path
doesn't start by the trash name. But there is also a trashed attribute,
and we can check both. If only one of those is true, we have an
inconsistency and we can hope that moving the file again to the trash
will help fixing it.