Fix installing app to a Swift container (#2701)
Swift is eventually consistent, and the stack may be fouled by this
behavior when installing an app.

To install an app to a Swift container, it creates a temporary
"directory", put the files from the tarball in this directory, and if
every thing is fine, copy the files to their final destination before
deleting them (moves don't exist in Swift). To get the list of files in
the temporary directory, the stack was calling the Swift container, with
a list+prefix operation. The response to this call can be not complete
when used without the `X-Newest` header (which was the case).

One way to fix the issue would have been to add this header, but the
Swift documentation recommends to avoid it, as it has some performance
costs. And, we can fix it by just keeping the list of uploaded files in
memory, which is better.