Fix echo on restore file deleted from a shared dir (#2731)
When a file was in a shared directory, was trashed, and is now restored,
the stack should be able to make the file available for all the members
of the sharing. It is not the case today, and this commit is not enough
to do so. But, at least, the file isn't moved again to the trash by the
sharing for the user who restored the file.