Fix opening a note from a revoked sharing
When a note has been shared and the sharing is revoked, all the members
of the sharing can continue to open the note on their instance. If
someone who is not the creator of the note opens it, they must be able
to edit it. It wasn't working when the cache expires, because of how the
steps were persisted (the sentinel step was not here).