Fix opening a note from a revoked sharing
When a note has been shared and the sharing is revoked, all the members
of the sharing can continue to open the note on their instance. If
someone who is not the creator of the note opens it, they must be able
to edit it. The changes were not persisted to the VFS because the
trigger was missing. The trigger is now recreated in this case when the
note is opened.