Add a fallback parameter for the BW icons route (#2791)
The pass clients show an icon for each domain that has a cipher. This
clients request the GET /bitwarden/icons/:domain/icon.png route of the
stack for this icon. This route has now a fallback parameter that allows
to return a 404 Not found error, instead of the default icon when no
favicon has been found for the domain. The clients can then use what
they want for the icon in that case.