Do not check the session on /move/initialize
The session cookie has a strict same-site policy, which means it won't
be sent for this request as the user is coming from a third site
(cozy-move). And we don't really need to check the session, as we
already check the token and we confirm the action by a mail.