Allow to run the delete accounts hook before an import
For some konnectors, there is a delete accounts hook that must be run
when an account is deleted to clear the account on the remote service.
When a Cozy is moved, on the target instance, the stack deletes all the
accounts before the import, but the konnectors were failing because the
instance was already blocked and the konnectors need to make some
requests to the stack.