Fix moving a file out and in of a shared folder
When a file is in a shared folder, and a member moves it out of this
sharing, the file is put in the trash for the other members. If someone
puts again this file in the shared folder, the safety rules forbid to
move it again in the shared folder for the other members. We have fixed
this issue but putting a copy of the file in trash (ande deleting the
file) to force a new ID.