Fix a sharing issue
When a file or directory is inside a folder shared between several
members, and a recipient moves this file or directory outside of the
sharing, the file or directory is dissociated on the owner's instance.
But the io.cozy.shared for it was deleted and, thus, the change was not
replicated to the other recipients. We can't keep the io.cozy.shared as
is, because it will block the file or directory to come back in the
sharing if the recipient moved it again to the shared folder. So, we
have introduced a new flag on the io.cozy.shared, dissociated, to
indicate this new state where a file or directory is seen as deleted
(and the change is replicated to all the recipients) but can still be
reuploaded inside the sharing.