Update the wording for the banner (#2978)
The message was abput checking the URL and the padlock before filling
the password. We remove the part about the padlock because there are
some contexts where the user is expected to give their password even if
the padlock is not here (webviews on android for example). And the
sentence was modified to emphasize it is a general tip for every time
the password must be given, not only for this time.