Fix view requests to CouchDB with keys (#2981)
When the stack send a request to CouchDB for a view with the keys
parameter, it is sent as a POST with the keys in the body. So, we should
not put the keys in the query-string, as it defeats the purpose of using
a POST in the first place (avoiding too long URLs).