Fix cleaning on uninstall konnectors (#2991)
When a konnector is uninstalled, the stack cleans the accounts. To do
that, it looks at the triggers, and when we had several triggers for the
same accounts (e.g. cron + webhook), it was trying to delete the
accounts several times, which failed. Now, the cleaning entries will
regroup the triggers with the same account to avoid this issue.