Fix the deletion of unused OAuth clients (#3036)
Some OAuth clients were used by sharings but their pending flag was not
removed, and after 1h, their clean-client trigger deletes them. Now, we
avoid setting the pending flag in first place when we know that a token
will be generated just after the client creation, which avoids this
issue.