Use the bitwarden settings ID for userId
For the bitwarden hub, let's use the identifier of the bitwarden
settings document as the userId for the notification messages. We don't
really care for the moment, but the userId will also be needed later for
sharing ciphers, and using the bitwarden settings ID will make things a
bit easier.