Fix userId for bitwarden clients
The bitwarden clients are parsing their access token as a JWT, and use
the subject field as a userId. This userId must match the one sent in
events from the bitwarden hub, and is also used to compute the
fingerprint.

Until this commit, the access token from the stack were saving the OAuth
client ID in the subject field of access token and refresh token.

For sharing ciphers between Cozy instances, we need to stable
fingerprints in bitwarden clients, and we had to introduce a hack for
their access tokens. The client ID is now saved in a new field of the
JWT (client_id), and the subject is the instance identifier. And we
overwrite the subject just after parsing for those tokens.

Access tokens from other clients, and all refresh tokens will continue
to work as before. The access token from bitwarden clients are modified
just after being parsed, so that they look like other access token in
stack code (and they will have a subject that is suitable as userId
client-side).

We also have to introduce a temporary work-around to force bitwarden
clients to refresh their token, and use the new format, to avoid
breaking the bitwarden hub. I hope it will work, but more tests are
needed with the clients to see how they manage a change of userId on a
refresh.