Fix bitwarden sync when updating a shared cipher
The bitwarden clients look at the revision date before doing a sync.
This date is updated when a cipher is created/updated/deleted on the
local instance, but wasn't updated when the change was coming from a
sharing. A call to update this date has been added to the sharing
mechanisms to fix this issue.