Fix deletion of shared documents
For io.cozy.files, the sharing was cheating a bit on the CouchDB
replication protocol to avoid long list of revisions in _revs_diff. For
other documents, we need the whole list to make the correct call to
bulk_docs with new_edits: false. It is particulary important for
deletions, as CouchDB will see the documents as a conflict fixed and not
a deletion if we do the bulk_docs update with only the last revision.