Fix removing a cipher from a shared organization
When a cipher is in a shared organization, and is moved to another
organization, the local stack sees an update, but the sharing
replication protocol must delete the cipher for the recipients. It is
done by marking the io.cozy.shared as removed in the share-track worker.
This clue is later interpreted by the worker-replicate that can send the
cipher document as deleted in the replication.