Allow to redirect to an app from login (#3178)
When using the redirect parameter of the /auth/login route, a full URL
was expected. But it can be complicated to know if the stack is in flat
or nested subdomains. It is now possible to use a short version of the
redirect parameter (like store/#/discover) to let the stack build the
correct host.