Add a trigger for clean-old-trashed if needed
When a file or directory is put in the trash, we look if a delay has
been configured for this instance context. If it is the case, we look if
a trigger is set up for clean-old-trashed, and create it if it missing.
It is a @cron trigger configured to run daily.