Fix HTTP caching of the webmanifest
This bug is similar to https://github.com/cozy/cozy-stack/pull/3285.
This time, the webmanifest was also accessed from the login page, where
there is no CORS, and the Vary: Origin header was added only when CORS.
So, the webmanifest was cached with no Vary: Origin header and no CORS
headers, and when the home wanted to serve the same asset, we had a CORS
error. This fix is quite simple: just force Vary: Origin, even if not
used with CORS.