Make session for OAuth danse expires after 5 minutes
We don't want the normal session for the OAuth danse, as it can lead to
weird behaviors. For example, if a user disconnects from a mobile app,
and tries again to reconnect to the same instance, we want them to log
in again (else, they might think that the disconnect doesn't clean the
data but it is just the InPageBrowser that has kept the session).