Forward warning for unoptimal index (#3430)
With CouchDB 3+, when an index can be used for a _find request, but
CouchDB still need to look to many documents that will not appear in
the response, a warning is emitted. This warning is now forwarded by the
stack to the client. The warning for no index at all is still
transformed to an error by the stack.